Fix string lexing for python
This fixes bugs with non-escaped quotes within strings (because
they don't need to be escaped if they are opposite the quoting type).
This is extremely common in python, and breaks things like '"'.

Also removes a bunch of useless lexing for escape characters.
Unlike vim, we aren't styling them a different color so theres not
a point in lexing them specifically.

In order to be fault-tolerant we just accept any character between
the strings, regardless of what python says is okay.