Add function to compose highlighters
This functionality will be essential if we ever want
to compose multiple modes which have syntax highlighting.

An obvious example would be adding a mode for rainbow parens
on top of a mode for a language. To do so we need to compose their
highlighters, because a mode can only have a single highlighter (and
having indefinitely many would cause issues with losing type
information, because their syntax type is different).

I want to formalize the categorical construction here in order
to sound smart, but I'm hilariously bad at that. Its probably
a bifunctor or something from ExtHL x ExtHL -> ExtHL.